pyproject.toml
[project]
name = "ai-host-agent"
version = "0.1.0"
description = "Voice-based restaurant reservation agent with outbound confirmation calls"
requires-python = ">=3.11"

dependencies = [
    # Web framework
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.30.0",
    "python-multipart>=0.0.9",
    # Data validation + settings
    "pydantic>=2.7.0",
    "pydantic-settings>=2.3.0",
    # HTTP client
    "httpx>=0.27.0",
    "aiofiles>=23.2.1",
    # LLM + agent framework
    "anthropic>=0.29.0",
    "langgraph>=0.2.0",
    "langgraph-checkpoint-sqlite>=2.0.0",
    "langchain-anthropic>=0.1.0",
    # Voice: STT + TTS
    "openai>=1.35.0",          # Whisper STT
    "elevenlabs>=1.0.0",       # ElevenLabs TTS
    # Telephony
    "twilio>=9.0.0",
    # Persistence
    "asyncpg>=0.29.0",
    "psycopg2-binary>=2.9.0",
    "redis>=5.0.0",
    # Vector store
    "lancedb>=0.10.0",
    "pyarrow>=15.0.0",
    # Retry / resilience
    "tenacity>=8.3.0",
    # Logging
    "structlog>=24.2.0",
    # Auth
    "bcrypt>=4.1.0",
    # Config
    "pyyaml>=6.0.1",
    # JWT
    "python-jose[cryptography]>=3.3.0",
    "python-dotenv>=1.0.1",
    # Scheduling (confirmation calls)
    "apscheduler>=3.10.4",
    # Observability
    "langfuse>=2.40.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=5.0.0",
    "respx>=0.21.0",
    "factory-boy>=3.3.0",
    "httpx>=0.27.0",
    "ruff>=0.4.0",
    "mypy>=1.10.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.mypy]
python_version = "3.11"
strict = true
ignore_missing_imports = true
